# specs/001-prompt-asset-hub/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Prompt Asset Hub API
  version: 0.1.0
  description: "本機單人模式 API（REST）"
servers:
  - url: http://localhost:3001
paths:
...
  /api/projects:
    get:
      summary: 列出專案（支援篩選/搜尋）
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Project' } } } } }
    post:
      summary: 新增專案
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectInput' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
  /api/prompts:
    get:
      summary: 列出提示詞（支援篩選/搜尋/看板）
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: priority
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Prompt' } } } } }
    post:
      summary: 新增提示詞
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PromptInput' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Prompt' } } } }
  /api/prompts/{id}:
    get:
      summary: 取得提示詞詳情
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Prompt' } } } }
    put:
      summary: 更新提示詞（正文 + 中繼資料）
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PromptInput' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Prompt' } } } }
    delete:
      summary: 刪除提示詞（移到回收站，可復原）
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TrashItem' } } } }
        '404': { description: Not Found }
  /api/projects/{id}:
    delete:
      summary: 刪除專案（移到回收站，可復原）
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TrashItem' } } } }
        '404': { description: Not Found }
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        summary: { type: string }
        status: { type: string }
        tags: { type: array, items: { type: string } }
        archived: { type: boolean }
        updatedAt: { type: string, format: date-time }
    ProjectInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        summary: { type: string }
        status: { type: string }
        tags: { type: array, items: { type: string } }
    Prompt:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        title: { type: string }
        status: { type: string }
        priority: { type: string }
        tags: { type: array, items: { type: string } }
        body: { type: string }
        notes: { type: string }
        attachments: { type: array, items: { $ref: '#/components/schemas/Attachment' } }
        updatedAt: { type: string, format: date-time }
    PromptInput:
      type: object
      required: [title]
      properties:
        projectId: { type: string }
        title: { type: string }
        status: { type: string }
        priority: { type: string }
        tags: { type: array, items: { type: string } }
        body: { type: string }
        notes: { type: string }
    Attachment:
      type: object
      properties:
        id: { type: string }
        filename: { type: string }
        mimeType: { type: string }
        storagePath: { type: string }
...
    TrashItem:
      type: object
      properties:
        trashId: { type: string }
        entityType: { type: string, enum: [project, prompt, inboxItem, snippet, unknown] }
        entityId: { type: string }
        titleSnapshot: { type: string }
        deletedAt: { type: string, format: date-time }
        purgeAfter: { type: string, format: date-time }
        originalRelativePath: { type: string }
        attachmentsMoved: { type: boolean }
...